pyproject.toml
[tool.poetry]

name = "friday"
version = "1.0.0"
description = "Keys AI Bot (Friday)"
authors = ["Keys (Shangar) <shangar202000@gmail.com>"]
readme = "README.md"
license = "LICENSE"
repository = "https://github.com/shangar-t-a/keys-ai-bot-friday/"

[tool.poetry.urls]
changelog = "CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
google-generativeai = "^0.7.2"
colorlog = "^6.8.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
